--- .gitignore
@@ -1,2 +1,0 @@
-./train/models/*
-./data/raw/*
